add util to format error to object
