+++ examples/deps.ts
@@ -1,3 +1,3 @@
 export { config } from 'https://deno.land/x/dotenv/mod.ts';
-export { CBFetch, FetchError } from '../mod.ts';
+export { CBFetch, fetchErrUtil } from '../mod.ts';
 export type { CBAccessSetupModel } from '../mod.ts';
